Fix the format of number in GUI
Previously I used rvalue.magnitude to only show the data magnitude and used rvalue.units to show units. But there is a problem that rvalue.magnitude ignored the format like 8.4f in the tango attribute. So I changed back to use rvalue to display both the magnitude and units in one widget